Re-enable the --self-contained option for HTML output
This is really quite nice to have enabled since it allows passing around
the document as a single file.  The option used to be present before
"Fix HTML template styling" (dc6ca66), when it was removed amongst other
changes that seemed unrelated.

The custom template must handle MathJax inclusion itself if it's needed,
since as the Pandoc --self-contained documentation notes, inlining the
MathJax.js script doesn't work.  To prevent inlining, the template
includes a data-external="1" attribute on the <script>.  Ideally Pandoc
itself would start including that attribute for MathJax, but until then
(if it ever comes to pass), we can work around it.